Restructure statistics related files
