+++ tests/unit_tests/test_qvalues_storey.py
@@ -3,7 +3,7 @@ import json
 import numpy as np
 from pytest import approx
 
-from mokapot.qvalues_storey import empirical_pvalues, estimate_pi0, qvalues
+from mokapot.stats.qvalues_storey import empirical_pvalues, estimate_pi0, qvalues
 
 
 def test_empirical_pvalues():
